backend/requirements.txt
django>=2.0.8
psycopg2-binary
djangorestframework
Pillow
djangorestframework-camel-case
graphene-django>=2.0
django-imagekit
